requirements.txt
Django==1.9
django-foundation==0.1
djangorestframework==3.3.2
gunicorn==19.4.1
psycopg2==2.6.1
wheel==0.26.0
